requirements.txt
streamlit
yfinance
beautifulsoup4
requests
google-genai
pandas
plotly
python-dotenv
numpy
